# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local microservices development environment API",
        "title": "DevArch API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8550",
    "basePath": "/api/v1",
    "paths": {
        "/auth/validate": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Validate API key",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/auth/ws-token": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Generate WebSocket authentication token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/stacks/{name}/apply": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Apply a generated plan to deploy stack changes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stacks"
                ],
                "summary": "Apply stack deployment plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stack name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Apply request with plan token",
                        "name": "apply",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.applyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ActionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
        "/stacks/{name}/compose": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stacks"
                ],
                "summary": "Get stack compose YAML",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Stack name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_priz_devarch-api_internal_api_respond.ActionResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "output": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "warnings": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_priz_devarch-api_internal_api_respond.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_priz_devarch-api_internal_api_respond.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_api_respond.ErrorDetail"
                }
            }
        },
        "github_com_priz_devarch-api_internal_api_respond.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
        "github_com_priz_devarch-api_internal_lock.InstLock": {
            "type": "object",
            "properties": {
                "image_digest": {
                    "type": "string"
                },
                "image_tag": {
                    "type": "string"
                },
                "resolved_ports": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "template": {
                    "type": "string"
                },
                "template_hash": {
                    "type": "string"
                }
            }
        },
        "github_com_priz_devarch-api_internal_lock.LockFile": {
            "type": "object",
            "properties": {
                "config_hash": {
                    "type": "string"
                },
                "generated_at": {
                    "type": "string"
                },
                "instances": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_priz_devarch-api_internal_lock.InstLock"
                    }
                },
                "stack": {
                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_lock.StackLock"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "github_com_priz_devarch-api_internal_lock.StackLock": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "network_id": {
                    "type": "string"
                },
                "network_name": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.applyRequest": {
            "type": "object",
            "properties": {
                "lock": {
                    "$ref": "#/definitions/github_com_priz_devarch-api_internal_lock.LockFile"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key authentication header",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}